Add a solver to the Countdown game (#373)
The solver is quite basic as it merely does a brute force
to find the best formula. The solver will only kick in if
nobody found a perfect formula.

The solver is disabled by default. You must pass an additional
"solver" parameter to the module's configuration that equals
to a boolean truth value to enable it.
Eg.  <param name="solver">yes</param>

The class used to actually solve the game can be changed
from the configuration file by setting the "solver_class"
parameter to the name of the class to use.
The default value is "Erebot_Module_Countdown_Solver".